requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
ca-certificates=2024.9.24=h06a4308_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.11=pypi_0
cython-bbox=0.1.5=pypi_0
faiss-cpu=1.8.0.post1=pypi_0
faiss-gpu=1.7.2=pypi_0
filelock=3.16.1=pypi_0
fonttools=4.55.2=pypi_0
fsspec=2024.10.0=pypi_0
gdown=5.2.0=pypi_0
grpcio=1.67.1=pypi_0
idna=3.10=pypi_0
imageio=2.36.0=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.7=pypi_0
lap=0.5.12=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
loguru=0.7.3=pypi_0
markdown=3.7=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.3=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.2=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py39h06a4308_0
protobuf=5.28.3=pypi_0
pycocotools=2.0.8=pypi_0
pyparsing=3.2.0=pypi_0
pysocks=1.7.1=pypi_0
python=3.9.20=he870216_1
python-dateutil=2.9.0.post0=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=pypi_0
setuptools=75.1.0=py39h06a4308_0
six=1.16.0=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.5.0=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.8.30=pypi_0
tk=8.6.14=h39e8969_0
torch=2.5.1=pypi_0
torchaudio=2.5.1=pypi_0
torchreid=0.2.5=pypi_0
torchvision=0.20.1=pypi_0
tqdm=4.67.0=pypi_0
trimesh=4.5.2=pypi_0
triton=3.1.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py39h06a4308_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
